Moved App to features directory
